rename index to server
